API: Remove const module and move get_namedtuple to util.
