+++ siriuspy/siriuspy/csdevice/pwrsupply.py
@@ -6,7 +6,7 @@ from siriuspy.csdevice.enumtypes import EnumTypes as _et
 from siriuspy.search import PSSearch as _PSSearch
 from siriuspy.search import MASearch as _MASearch
 from siriuspy.pwrsupply.siggen import DEFAULT_SIGGEN_CONFIG as _DEF_SIGG_CONF
-from siriuspy.csdevice.const import get_namedtuple as _get_namedtuple
+from siriuspy.util import get_namedtuple as _get_namedtuple
 
 # MIN_WFMSIZE = 2001
 MAX_WFMSIZE = 4000
