+++ siriuspy/siriuspy/util.py
@@ -14,6 +14,7 @@ import datetime as _datetime
 import epics as _epics
 import numpy as _np
 import sys as _sys
+from collections import namedtuple as _namedtuple
 
 from siriuspy import envars as _envars
 
@@ -325,3 +326,21 @@ def check_public_interface_namespace(namespace, valid_interface,
                 print('Missing symbol: ', name)
             return False
     return True
+
+
+def get_namedtuple(name, field_names, values=None):
+    """Return an instance of a namedtuple Class.
+
+    Inputs:
+        - name:  Defines the name of the Class (str).
+        - field_names:  Defines the field names of the Class (iterable).
+        - values (optional): Defines field values . If not given, the value of
+            each field will be its index in 'field_names' (iterable).
+
+    Raises ValueError if at least one of the field names are invalid.
+    Raises TypeError when len(values) != len(field_names)
+    """
+    if values is None:
+        values = range(len(field_names))
+    field_names = [f.replace(' ', '_') for f in field_names]
+    return _namedtuple(name, field_names)(*values)
